+++ R/import_export.R
@@ -11,15 +11,15 @@
 #' @param region Only used in xlsx import. Can either be an 'Excel' range like 'A1:BY27'
 #' or the name of a named region.
 #' @param data_frame A data frame to export.
-#' @param outfile Full file path with extension. Allowed extensions are ".csv" and ".xlsx".
-#' @param separator Only used in CSV-export. Defines the single character value separator.
-#' @param decimal Only used in CSV-export. Defines the single character decimal character.
+#' @param outfile Full file path with extension. Allowed extensions are ".csv", ".txt" and ".xlsx".
+#' @param separator Only used in CSV/TXT-export. Defines the single character value separator.
+#' @param decimal Only used in CSV/TXT-export. Defines the single character decimal character.
 #' @param var_names TRUE by default. Whether to export variable names or not.
 #'
 #' @details
 #' [import_data()] and [export_data()] are based on the 'SAS' procedures Proc Import and Proc Export,
 #' which provide a very straight forward syntax. While 'SAS' can import many different formats with
-#' these procedures, these 'R' versions concentrate on importing CSV and XLSX files.
+#' these procedures, these 'R' versions concentrate on importing CSV, TXT and XLSX files.
 #'
 #' The main goal here is to just provide as few as possible parameters to tackle most of the imports
 #' and exports. These error handling also tries to let an import and export happen, even though
@@ -39,11 +39,13 @@
 #'
 #' @examples
 #' # Example files
-#' csv_file  <- system.file("extdata", "qol_example_data_csv.csv",  package = "qol")
+#' csv_file  <- system.file("extdata", "qol_example_data_csv.csv",   package = "qol")
+#' txt_file  <- system.file("extdata", "qol_example_data_txt.txt",   package = "qol")
 #' xlsx_file <- system.file("extdata", "qol_example_data_xlsx.xlsx", package = "qol")
 #'
 #' # Import: Provide full file path
 #' my_csv  <- import_data(csv_file)
+#' my_csv  <- import_data(txt_file)
 #' my_xlsx <- import_data(xlsx_file)
 #'
 #' # Import specific regions
@@ -104,8 +106,8 @@ import_data <- function(infile,
             return(invisible(NULL))
         }
 
-        if (!extension %in% c("csv", "xlsx")){
-            message(" X ERROR: Only 'csv' or 'xlsx' are allowed as file extensions in the <infile>.\n",
+        if (!extension %in% c("csv", "txt", "xlsx")){
+            message(" X ERROR: Only 'csv', 'txt' or 'xlsx' are allowed as file extensions in the <infile>.\n",
                     "          Import will be aborted.")
 
             return(invisible(NULL))
@@ -116,7 +118,7 @@ import_data <- function(infile,
     # Import
     ###########################################################################
 
-    if (extension == "csv"){
+    if (extension %in% c("csv", "txt")){
         #++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
         # Separator
         #++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
@@ -262,7 +264,7 @@ import_data <- function(infile,
 #'
 #' @examples
 #' # Import multiple files at once
-#' all_files <- import_multi(c(csv_file, xlsx_file))
+#' all_files <- import_multi(c(csv_file, txt_file, xlsx_file))
 #'
 #' @rdname import_export
 #'
@@ -285,7 +287,7 @@ import_multi <- function(file_list,
         extension <- tolower(tools::file_ext(infile))
 
         # For CSV files just do a simple import
-        if (extension == "csv"){
+        if (extension %in% c("csv", "txt")){
             message(" > Importing: ", infile)
 
             result_list[[filename]] <- suppressMessages(
@@ -347,10 +349,12 @@ import_multi <- function(file_list,
 #'
 #' # Example export file paths
 #' export_csv  <- tempfile(fileext = ".csv")
+#' export_txt  <- tempfile(fileext = ".txt")
 #' export_xlsx <- tempfile(fileext = ".xlsx")
 #'
 #' # Export: Provide full file path
 #' my_data |> export_data(export_csv)
+#' my_data |> export_data(export_txt)
 #' my_data |> export_data(export_xlsx)
 #'
 #' @rdname import_export
@@ -399,14 +403,14 @@ export_data <- function(data_frame,
         extension <- "csv"
     }
 
-    if (!extension %in% c("csv", "xlsx")){
-        message(" ! WARNING: Only 'csv' or 'xlsx' are allowed as file extensions in the <outfile>. 'csv' will be used.")
+    if (!extension %in% c("csv", "txt", "xlsx")){
+        message(" ! WARNING: Only 'csv', 'txt' or 'xlsx' are allowed as file extensions in the <outfile>. 'csv' will be used.")
 
         outfile <- sub(extension, "csv", outfile, ignore.case = TRUE)
         extension <- "csv"
     }
 
-    if (extension == "csv"){
+    if (extension %in% c("csv", "txt")){
         #++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
         # Separator
         #++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
@@ -490,8 +494,8 @@ export_data <- function(data_frame,
 #'
 #' @examples
 #' # Example data frame list
-#' my_list <- list(first  = dummy_data(10),
-#'                 second = dummy_data(10))
+#' my_list <- list(first      = dummy_data(10),
+#'                 second.txt = dummy_data(10))
 #'
 #' # Export multiple data frames into one xlsx file
 #' # with multiple sheets
@@ -500,7 +504,7 @@ export_data <- function(data_frame,
 #' # Export multiple data frames into multiple xlsx files
 #' export_multi(my_list, tempdir(), into_sheets = FALSE)
 #'
-#' # Export multiple data frames into multiple csv files
+#' # Export multiple data frames into multiple csv/txt files
 #' export_multi(my_list, tempdir(), separator = ";")
 #'
 #' # Manual cleanup for example
@@ -508,7 +512,7 @@ export_data <- function(data_frame,
 #' file2 <- file.path(tempdir(), "first.xlsx")
 #' file3 <- file.path(tempdir(), "second.xlsx")
 #' file4 <- file.path(tempdir(), "first.csv")
-#' file5 <- file.path(tempdir(), "second.csv")
+#' file5 <- file.path(tempdir(), "second.txt")
 #'
 #' unlink(c(export_csv, export_xlsx,
 #'          file1, file2, file3, file4, file5))
@@ -537,8 +541,18 @@ export_multi <- function(file_list,
 
         # For CSV files just do a simple export
         if (!is.null(separator)){
+            extension <- tolower(tools::file_ext(filename))
+
+            if (!extension %in% c("csv", "txt")){
+                if (!extension == ""){
+                    message(" ! WARNING: Only 'csv', 'txt' or 'xlsx' are allowed as file extensions. 'csv' will be used.")
+                }
+
+                filename <- paste0(filename, ".csv")
+            }
+
             suppressMessages(export_data(data_frame,
-                                         outfile   = paste0(out_path, "/", filename, ".csv"),
+                                         outfile   = paste0(out_path, "/", filename),
                                          separator = separator,
                                          decimal   = decimal,
                                          var_names = var_names))